chore: moved json converter logic into json renderer Co-authored-by: lottejd@gmail.com
