+++ src/c-sharp/Infra/JsonRenderer.cs
@@ -8,8 +8,44 @@ namespace Archlens.Infra;
 
 public sealed class JsonRenderer : IRenderer
 {
-    public Task<string> RenderGraphAsync(DependencyGraph graph, Options options, CancellationToken ct = default)
+    public string RenderGraphAsync(DependencyGraph graph, Options options, CancellationToken ct = default)
     {
-        throw new System.NotImplementedException();
+        var packagestr = "";
+        for (int i = 0; i < graph.packages().Count; i++)
+        {
+            var package = graph.packages()[i];
+
+            if (packagestr.Contains(package)) continue;
+
+            var comma = "";
+            if (i < graph.packages().Count - 2) comma = ",";
+
+            packagestr +=
+                $$"""
+                
+                {
+                    "name": "{{package}}",
+                    "state": "NEUTRAL"
+                }{{comma}}
+
+            """;
+        }
+
+        var str =
+        $$"""
+        {
+            "title": "{{graph.Name}}",
+            "packages": [
+                {{packagestr}}
+            ],
+
+            "edges": [
+            {{graph.ToJson()}}
+
+            ]
+        }
+        """;
+        return str;
     }
+
 }
